adding sync
